add __init__.py to tests dir move train.py and predict.py
